requirements.txt
 
openai
python-dotenv
faiss-cpu
sentence-transformers
PyPDF2
python-docx
openpyxl
uvicorn
fastapi
pytest